Added separate file for Header and Footer
